Add new filed `attributeOwnerIdBeforeInline` to IrAttributeContainer
This field is null by default, and it is set to current
`attributeOwnerId` when declaration was inlined and must be
generated with new name. In its turn `attributeOwnerId` must be set to
`this` to show that this is actually a new declaration, not a copy.

(cherry picked from commit 60702f0f02ebb8039f778852ebeb04c4a8e262bc)