[IR] Support actualizing expect companion members to static Java members
We do this in two steps. First we adapt the matching logic to also
search for static members when the declaration is a companion member.
Second, we remove companion dispatch receivers when we encounter a call
to a member of an actual (through typealias) Java class. The same
applies to callable references.

#KT-57661 Fixed