2/2 MUST_BE_INITIALIZED: take into account containingDeclaration's modality
This commit is important in scope of KT-57553. It makes the migration
more smooth.

The previous implementation was too restrictive. It didn't consider
"effectively final" case

Because of that restrictive behaviour the previous implementation had
too many introduced compilation errors in intellij project (93
compilation errors are introduced by this commit). After this commit is
applied, number of compilation errors became 3 in intellij project.

Other related tests:
- testUninitializedOrReassignedVariables