squash! Update KClass.isData KDoc to include info about data objects
`data object` is introduced only in 1.9, but it looks like we got
correct `KClass.isData` behaviour for "free" (`isData` returns `true`
for `data object`s even if kotlin-reflect version is old (1.2.0)).
That's why this behaviour was never covered with tests.

This commit updates the documentation and covers new behaviour with
tests.