KTIJ-25461 [Analysis API] Unwrap substitution overrides for FirNamedReference in KtFirExpressionTypeProvider
This is done to avoid leaking potential CapturedType(*) outside of
the Analysis API, because such types cannot be processed by the
UAST and Light Classes (in particular by `AbstractTypeMapper`,
which maps Kotlin types to JVM types).

`FirNamedReference` might appear when resolving method references (like
`foo::bar`), but also when IJ Platform tries to resolve other parts of
the Kotlin PSI, notably a `KtNameReferenceExpression` in a
function call (`bar` in `foo.bar(baz)` expression).

N.B. It's not clear whether KtFirExpressionTypeProvider should provide
the type for such expressions at all - it is possible that this code
should be rewritten to return null instead of trying to return some
type (see KT-59077).

^KTIJ-25461 Fixed